+++ js/ui/tag_box.js
@@ -18,7 +18,7 @@ import { addNamespace, normalizeKeyName } from '../events/utils/index';
 import { name as clickEvent } from '../events/click';
 import caret from './text_box/utils.caret';
 import { normalizeLoadResult } from '../data/data_source/utils';
-import { getScrollRtlBehavior } from '../core/utils/scrollRtlBehavior';
+import getScrollRtlBehavior from '../core/utils/scroll_rtl_behavior';
 
 import SelectBox from './select_box';
 import { BindableTemplate } from '../core/templates/bindable_template';
